Refactor: Extract FillingSQLWithParamMarker
